Refactor: 게시판 페이지네이션 컴포넌트
