+++ src/Pages/Video/AllVideo.js
@@ -16,15 +16,15 @@ import {
 } from "../../Style/Video";
 import TV from '../../Assets/TV.png';
 import {FontAwesomeIcon} from "@fortawesome/react-fontawesome";
-import {faCaretLeft, faCaretRight, faHeart} from "@fortawesome/free-solid-svg-icons";
+import {faHeart} from "@fortawesome/free-solid-svg-icons";
 import StyledBtn from "../../Style/StyledBtn";
 import preURL from "../../preURL/preURL";
 import axios from "axios";
 import PlayListIcon from "../../Assets/Playlist_mini.png";
 import {useNavigate} from "react-router-dom";
-import {PageNum, Pages, Pagination} from "../../Style/Community";
 import Add_New_Video from "../../Assets/Add_new_video.png";
 import useInput from "../../Hooks/useInput";
+import Pagination from "../../Components/Pagination";
 
 const AllVideo = () => {
   const navigate = useNavigate();
@@ -44,8 +44,6 @@ const AllVideo = () => {
   const [totalPageCount, setTotalPageCount] = useState(0);  // 총 페이지 수
   const [pages, setPages] = useState([1,2,3,4,5]);
   const [page, setPage] = useState(0);  // 현재 페이지
-  const [showGoLeftPages, setShowGoLeftPages] = useState(false);
-  const [showGoRightPages, setShowGoRightPages] = useState(true);
   const [sort, setSort] = useState(""); // 좋아요 순이면 -> likeCount,DESC
   const [playlistToggleDisplay, setPlaylistToggleDisplay] = useState(false);  // 플레이리스트 모달창 보이기
   const [playlistList, setPlaylistList] = useState([]); // 받아온 내 플레이리스트 목록
@@ -201,47 +199,6 @@ const AllVideo = () => {
     setSort("likeCount,DESC");
   };
 
-  // 페이지 클릭
-  const onClickPage = (e) => {
-    setPage(e.target.innerHTML-1);
-    console.log("페이지 클릭", e.target.innerHTML);
-  };
-
-  // 페이지 넘기기
-  const onClickNextPages = () => {
-    console.log("페이지 넘기기");
-    console.log("페이지:" + pages);
-    if(pages.length < 5) return;
-    let list = [];
-    for(let i=0; i<5; i++){
-      list[i] = pages[i]+5;
-      if(list[i] >= totalPageCount) {
-        setShowGoRightPages(false);
-        break;
-      }
-    }
-    setPages(list);
-    setShowGoLeftPages(true);
-  };
-
-  //페이지 돌아가기
-  const onClickPreviousPages = () => {
-    console.log("페이지 돌아가기");
-    console.log("페이지: " + pages);
-    if(pages.length < 5){
-      for(let i=0; i<5; i++) pages[i] = pages[0]+i;
-    }
-    let list = [];
-    for(let i=4; i>=0; i--){
-      list[i] = pages[i]-5;
-      if(list[i] <= 1) {
-        setShowGoLeftPages(false);
-        break;
-      }
-    }
-    setPages(list);
-    setShowGoRightPages(true);
-  }
 
   // 각 비디오
   const OneVideo = (video) => {
@@ -316,14 +273,6 @@ const AllVideo = () => {
     )
   };
 
-  // 페이지 번호
-  const showPages = pages.map((page) => {
-    return (
-        <PageNum style={{fontSize: "20px", padding: "10.5px"}} onClick={onClickPage}>
-          {page}
-        </PageNum>
-    )
-  });
 
   return (
       <div>
@@ -370,33 +319,8 @@ const AllVideo = () => {
               좋아요순
             </StyledBtn>
           </SortBox>
-          <Pagination>
-            {showGoLeftPages &&
-                <StyledBtn id="previous-page" onClick={onClickPreviousPages}>
-                  <FontAwesomeIcon
-                      icon={faCaretLeft}
-                      style={{
-                        fontSize: "20px",
-                        color: "#9C9C9C",
-                        marginLeft: "10.5px",
-                      }}
-                  />
-                </StyledBtn>
-            }
-            <Pages>{showPages}</Pages>
-            {showGoRightPages &&
-                <StyledBtn id="next-page" onClick={onClickNextPages}>
-                  <FontAwesomeIcon
-                      icon={faCaretRight}
-                      style={{
-                        fontSize: "20px",
-                        color: "#9C9C9C",
-                        marginLeft: "10.5px",
-                      }}
-                  />
-                </StyledBtn>
-            }
-          </Pagination>
+          {/*페이지네이션*/}
+          <Pagination pages={pages} setPages={setPages} setPage={setPage} totalPageCount={totalPageCount} />
         </Wrapper>
       </div>
   )
